requirements.txt
click==8.1.3
Flask==2.2.5
Flask-Cors==3.0.10
gunicorn==20.1.0
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.6
six==1.16.0
sqlparse==0.4.4
tabulate==0.9.0
typing-extensions==4.5.0
Werkzeug==2.2.3
yoyo-migrations==8.2.0
zipp==3.15.0
